requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Flask-Migrate==2.2.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
gunicorn==19.9.0